src/models/blip/requirements.txt
# BLIP specific requirements
# Optimized for Docker deployment

# Core dependencies
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
opencv-python>=4.8.0
Pillow>=9.5.0

# BLIP/Transformers
transformers>=4.30.0
accelerate>=0.20.0

# API dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
pydantic>=2.0.0

# Utilities
requests>=2.31.0
aiofiles>=23.0.0
python-multipart>=0.0.6

# Barcode detection
pyzbar>=0.1.9
opencv-contrib-python>=4.8.0






